+++ src/parts/slot.js
@@ -95,7 +95,8 @@ export function attachSlot(slotName, label, node) {
                 ctx.build(n.placeholder);
             });
             ctx.write(true, '}');
-        }
+        } else if(hasDynProps && !this.config.immutable) ctx.write(`, 0`)
+        if(hasDynProps && !this.config.immutable) ctx.write(`, $runtime.$$compareDeep`)
         ctx.write(');\n');
     });
 };